+++ RELEASES.md
@@ -9,6 +9,8 @@ Some of the changes done in this version is:
 - Fixed encoding problem when opening a CSV-file that was edited in Excel using ISO-8859-1 aka Windows-1252 aka CP1252
 - Added keybord shortcuts for all the menu options
 - Improved user experience when panning the Timeline using the scroll-mouse-button
+- Fixed "ghost lines" in CSV where Excel is not removing the delimiter when removing a line
+- Fixed the blank property that originated from the 5th column in the CSV file representing the JSON data
 
 ## 1.2.0
 2022-08-07